tests: move the fake mcv interface to a different file
This is to prepare for individual controller packages.
